BinanceStream/requirements.txt
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
matplotlib==3.8.2
ipython==8.21.0
Jinja2==3.1.3
websocket-client==1.7.0
wheel==0.43.0
